Add mixins for changeset properties :visibility and :apply_remote_metadata
