+++ spec/change_sets/ephemera_folder_change_set_spec.rb
@@ -98,10 +98,4 @@ RSpec.describe EphemeraFolderChangeSet do
       expect(change_set.subject).to be_empty
     end
   end
-
-  describe "#apply_remote_metadata?" do
-    it "overrides to be false" do
-      expect(change_set.apply_remote_metadata?).to be false
-    end
-  end
 end
